Revert "reformat settings menu"
This reverts commit a9b7c0fdc22db04cac590a6c4d525e8e63c4d961.
