+++ build.gradle
@@ -16,26 +16,23 @@ repositories {
 
 dependencies {
     //compile 'com.android.support:support-v4:23.0.1'                       // support library (Drawers) // pulled by MDDialogs
+    compile 'pl.droidsonroids.gif:android-gif-drawable:1.1.7'
     // animated gif drawables
     //compile 'com.android.support:appcompat-v7:23.0.1'
     // support library (material design)  // pulled by MDDialogs
+    compile 'com.afollestad:material-dialogs:0.7.9.1'
     // material design for dialogs
     //compile 'org.jsoup:jsoup:1.8.3'
+    compile 'org.mozilla:rhino:1.7.6'
     // JS for CloudFlare solving
+    compile 'com.squareup.okhttp:okhttp:2.5.0'
     // HTTP client
+    compile 'com.google.code.gson:gson:2.3.1'
     // For Prostopleer Json (de)serializing
+    compile 'com.android.support:design:23.0.1'
     compile fileTree(include: '*.jar', dir: 'libs')
     // apache html entity + custom JSoup
-    compile 'pl.droidsonroids.gif:android-gif-drawable:1.1.7'
-    compile 'com.afollestad:material-dialogs:0.7.9.1'
-    compile 'org.mozilla:rhino:1.7.6'
-    compile 'com.squareup.okhttp:okhttp:2.5.0'
-    compile 'com.google.code.gson:gson:2.3.1'
-    compile 'com.android.support:design:23.0.1'
-    compile 'com.android.support:support-v4:23.0.1'
 }
-
-
 
 android {
     compileSdkVersion 23
@@ -47,13 +44,13 @@ android {
         versionCode 230
         versionName "2.3.0"
     }
-
+    
     try {
         def proc = 'git rev-parse --short HEAD'.execute()
         def gitOutput = new StringBuilder('-')
         proc.waitForProcessOutput(gitOutput, System.err)
         def hash = gitOutput.toString().replace('\n', '').replace('\r', '')
-        if (!hash.empty) {
+        if(!hash.empty) {
             defaultConfig.versionName += hash
         }
     } catch (ignored) {
