--- src/main/res/xml/pref_client_view.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
-    <ListPreference
-        android:entries="@array/preference_themes_titles"
-        android:entryValues="@array/preference_themes_values"
-        android:key="app.theme"
-        android:summary="@string/app_theme_summary"
-        android:title="@string/app_theme"
-        android:negativeButtonText="@null"
-        android:positiveButtonText="@null" />
-    <SwitchPreference
-        android:defaultValue="false"
-        android:key="images.autoload"
-        android:summary="@string/images_autoload_summary"
-        android:title="@string/images_autoload" />
-
-    <EditTextPreference
-        android:defaultValue="12"
-        android:digits="1234567890"
-        android:key="webview.font.size"
-        android:summary="@string/webview_font_size_sum"
-        android:title="@string/webview_font_size" />
-    <SwitchPreference
-        android:defaultValue="false"
-        android:key="use.text.links"
-        android:summary="@string/use_text_links_sum"
-        android:title="@string/use_text_links" />
-</PreferenceScreen>
