--- src/main/res/xml/pref_headers.xml
@@ -1,17 +1,0 @@
-<preference-headers xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <!-- These settings headers are only used on tablets. -->
-
-    <header
-        android:fragment="adonai.diary_browser.preferences.SettingsActivity$ClientViewFragment"
-        android:title="@string/pref_header_client_view" />
-
-    <header
-        android:fragment="adonai.diary_browser.preferences.SettingsActivity$DiaryFragment"
-        android:title="@string/pref_header_diary" />
-
-    <header
-        android:fragment="adonai.diary_browser.preferences.SettingsActivity$AppSettingsFragment"
-        android:title="@string/pref_header_app_settings" />
-
-</preference-headers>
